+++ src/components/topbar/WorkflowTabs.vue
@@ -85,8 +85,8 @@ import { useWorkflowService } from '@/platform/workflow/core/services/workflowSe
 import {
   ComfyWorkflow,
   useWorkflowBookmarkStore
-} from '@/platform/workflow/ui/stores/workflowStore'
-import { useWorkflowStore } from '@/platform/workflow/ui/stores/workflowStore'
+} from '@/platform/workflow/management/stores/workflowStore'
+import { useWorkflowStore } from '@/platform/workflow/management/stores/workflowStore'
 import { useCommandStore } from '@/stores/commandStore'
 import { useSettingStore } from '@/stores/settingStore'
 import { useWorkspaceStore } from '@/stores/workspaceStore'
