+++ src/composables/graph/useCanvasRefresh.ts
@@ -1,5 +1,5 @@
 // call nextTick on all changeTracker
-import { useWorkflowStore } from '@/platform/workflow/ui/stores/workflowStore'
+import { useWorkflowStore } from '@/platform/workflow/management/stores/workflowStore'
 import { useCanvasStore } from '@/renderer/core/canvas/canvasStore'
 
 /**
