+++ src/platform/workflow/management/stores/workflowStore.ts
@@ -4,9 +4,9 @@ import { type Raw, computed, markRaw, ref, shallowRef, watch } from 'vue'
 
 import { t } from '@/i18n'
 import type { LGraph, Subgraph } from '@/lib/litegraph/src/litegraph'
-import { useWorkflowThumbnail } from '@/platform/workflow/rendering/composables/useWorkflowThumbnail'
 import { ComfyWorkflowJSON } from '@/platform/workflow/validation/schemas/workflowSchema'
 import type { NodeId } from '@/platform/workflow/validation/schemas/workflowSchema'
+import { useWorkflowThumbnail } from '@/renderer/thumbnail/useWorkflowThumbnail'
 import { api } from '@/scripts/api'
 import { app as comfyApp } from '@/scripts/app'
 import { ChangeTracker } from '@/scripts/changeTracker'
