+++ src/platform/workflow/persistence/composables/useWorkflowAutoSave.ts
@@ -1,7 +1,7 @@
 import { computed, onUnmounted, watch } from 'vue'
 
 import { useWorkflowService } from '@/platform/workflow/core/services/workflowService'
-import { useWorkflowStore } from '@/platform/workflow/ui/stores/workflowStore'
+import { useWorkflowStore } from '@/platform/workflow/management/stores/workflowStore'
 import { api } from '@/scripts/api'
 import { useSettingStore } from '@/stores/settingStore'
